requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.17.0
textblob>=0.17.1
nltk>=3.8.1
together>=0.2.11
python-dotenv>=1.0.0
requests>=2.31.0